pyproject.toml
[tool.poetry]
name = "opendatacancer"
version = "0.1.0"
description = ""
authors = ["Francisco Schulz <22344801+fschlz@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pandas = "^1.3.3"
streamlit = "^0.89.0"
plotly = "^5.3.1"
PyYAML = "^6.0"
numpy = "^1.21.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.9b0"
mypy = "^0.910"
pylint = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
